Allow none for bindings
also rearranged some patches
